naming fix behaviour -> strategy
