+++ src/main/java/fivefishes/design/patterns/group/assignment/components/strategy/StrategyControlPanel.java
@@ -1,17 +1,17 @@
-package fivefishes.design.patterns.group.assignment.components.behaviour;
+package fivefishes.design.patterns.group.assignment.components.strategy;
 
 import fivefishes.design.patterns.group.assignment.components.DesignPatternControlPanel;
-import fivefishes.design.patterns.group.assignment.controllers.behaviour.LightBehaviourController;
+import fivefishes.design.patterns.group.assignment.controllers.strategy.LightBehaviourController;
 
-public class BehaviourControlPanel extends DesignPatternControlPanel {
+public class StrategyControlPanel extends DesignPatternControlPanel {
     private LightBehaviourController lightBehaviourController;
     private LanternLightComboBox lanternLightComboBox;
 
-    public BehaviourControlPanel(String title, LightBehaviourController lightBehaviourController) {
+    public StrategyControlPanel(String title, LightBehaviourController lightBehaviourController) {
         this(title, "", lightBehaviourController);
     }
 
-    public BehaviourControlPanel(String title, String description, LightBehaviourController lightBehaviourController) {
+    public StrategyControlPanel(String title, String description, LightBehaviourController lightBehaviourController) {
         super(title, description);
         this.lightBehaviourController = lightBehaviourController;
         this.lanternLightComboBox = new LanternLightComboBox(lightBehaviourController);
