+++ src/main/java/fivefishes/design/patterns/group/assignment/entities/strategy/Bright.java
@@ -1,0 +1,12 @@
+
+package fivefishes.design.patterns.group.assignment.entities.strategy;
+
+import fivefishes.design.patterns.group.assignment.interfaces.strategy.LightBehaviour;
+
+public class Bright implements LightBehaviour {
+
+    @Override
+    public float lightIntensity() {
+        return 0.5f;
+    }
+}
