+++ src/main/java/fivefishes/design/patterns/group/assignment/entities/strategy/Lantern.java
@@ -1,7 +1,7 @@
 
-package fivefishes.design.patterns.group.assignment.entities.behaviour;
+package fivefishes.design.patterns.group.assignment.entities.strategy;
 
-import fivefishes.design.patterns.group.assignment.interfaces.behaviour.LightBehaviour;
+import fivefishes.design.patterns.group.assignment.interfaces.strategy.LightBehaviour;
 import java.awt.image.BufferedImage;
 
 public class Lantern {
