+++ src/main/java/fivefishes/design/patterns/group/assignment/resources/strategy/green_lantern.png
